chatbot_forecast/requirements.txt
dash
dash-bootstrap-components
google-generativeai
python-dotenv
requests
pandas
isodate
